[PATCH] uml: 64-bit type cleanups
64-bit cleanliness - Fix the number of bits of the time_t field in the COW
header to be 32 and change an int to a longs.

+++ arch/um/include/kern_util.h
@@ -41,7 +41,7 @@ extern int is_tracing(void *task);
 extern int segv_syscall(void);
 extern void kern_finish_exec(void *task, int new_pid, unsigned long stack);
 extern int page_size(void);
-extern int page_mask(void);
+extern unsigned long page_mask(void);
 extern int need_finish_fork(void);
 extern void free_stack(unsigned long stack, int order);
 extern void add_input_request(int op, void (*proc)(int), void *arg);
